+++ sdk/formrecognizer/Azure.AI.FormRecognizer/src/FieldValue_internal.cs
@@ -8,6 +8,9 @@ namespace Azure.AI.FormRecognizer.Models
     [CodeGenModel("FieldValue")]
     internal partial class FieldValue_internal
     {
+        /// <summary>Integer value.</summary>
+        public long? ValueInteger { get; }
+
         internal FieldValue_internal(string value)
         {
             Type = FieldValueType.String;