+++ sdk/formrecognizer/Azure.AI.FormRecognizer/CHANGELOG.md
@@ -23,6 +23,7 @@
 - `RecognizeContentOperation` now returns a `FormPageCollection`.
 - `RecognizeReceiptsOperation` now returns a `RecognizedReceiptCollection`.
 - `RecognizeCustomFormsOperation` now returns a `RecognizedFormCollection`.
+- In preparation for service-side changes, `FieldValue.AsInt32` has been replaced by `FieldValue.AsInt64`, which returns a `long`.
 
 ### New Features